(feat) Added helm chart to deploy the application
This commit adds the helm chart to deploy the application. Helm chart dependencies were added for nginx ingress controller and kube-prometheus-stack which can be disabled based on Helm chart values.

The application code was modified to support database connection retries and prevent the pod from crashing on startup. An additional healthz endpoint was added to test connections to Redis and Postgres. This endpoint is used by the ReadinessProbe to check if the application pod can connect to its dependencies before marking the container as 'Ready'.

In order to pass the postgres password to the database and to the application, we are making use of the helm-secrets plugin which decrypts secrets on the fly during the helm release installation.